Refactor to disentangle field order and layout
