+++ src/CMakeLists.txt
@@ -1,3 +1,3 @@
 find_package(Clang REQUIRED)
-add_library(randstruct SHARED randstruct.cpp)
+add_library(randstruct SHARED plugin.cpp randstruct.cpp)
 target_include_directories(randstruct PUBLIC ${CLANG_INCLUDE_DIRS})
