+++ src/plugin.cpp
@@ -1,0 +1,64 @@
+#include <stack>
+
+#include "clang/AST/AST.h"
+#include "clang/AST/ASTConsumer.h"
+#include "clang/AST/RecursiveASTVisitor.h"
+#include "clang/Frontend/FrontendPluginRegistry.h"
+#include "clang/Sema/Sema.h"
+#include "llvm/Support/raw_ostream.h"
+
+#include "randstruct.h"
+
+namespace {
+class RandstructConsumer : public ASTConsumer {
+  CompilerInstance &Instance;
+
+public:
+  RandstructConsumer(CompilerInstance &Instance) : Instance(Instance) {
+    Instance.getASTContext().setExternalSource(
+        llvm::IntrusiveRefCntPtr<Randstruct>(new Randstruct(Instance)));
+  }
+};
+
+class RandstructAction : public PluginASTAction {
+protected:
+  std::unique_ptr<ASTConsumer> CreateASTConsumer(CompilerInstance &CI,
+                                                 llvm::StringRef) override {
+    return llvm::make_unique<RandstructConsumer>(CI);
+  }
+
+  bool ParseArgs(const CompilerInstance &CI,
+                 const std::vector<std::string> &args) override {
+    for (unsigned i = 0, e = args.size(); i != e; ++i) {
+      llvm::errs() << "Randstruct arg = " << args[i] << "\n";
+
+      // Example error handling.
+      DiagnosticsEngine &D = CI.getDiagnostics();
+      if (args[i] == "-an-error") {
+        unsigned DiagID = D.getCustomDiagID(DiagnosticsEngine::Error,
+                                            "invalid argument '%0'");
+        D.Report(DiagID) << args[i];
+        return false;
+      } else if (args[i] == "-parse-template") {
+        if (i + 1 >= e) {
+          D.Report(D.getCustomDiagID(DiagnosticsEngine::Error,
+                                     "missing -parse-template argument"));
+          return false;
+        }
+        ++i;
+      }
+    }
+    if (!args.empty() && args[0] == "help")
+      PrintHelp(llvm::errs());
+
+    return true;
+  }
+  void PrintHelp(llvm::raw_ostream &ros) {
+    ros << "Help for Randstruct plugin goes here\n";
+  }
+};
+
+} // namespace
+
+static FrontendPluginRegistry::Add<RandstructAction>
+    X("randstruct", "randomizes the layout of structures");
